% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{convert_array_to_raster}
\alias{convert_array_to_raster}
\title{RGB or RGBA Array to Raster}
\usage{
convert_array_to_raster(arr)
}
\arguments{
\item{arr}{array}
}
\value{
raster
}
\description{
RGB or RGBA Array to Raster
}
