% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{convert_raster_to_array}
\alias{convert_raster_to_array}
\title{Raster to RGBA array}
\usage{
convert_raster_to_array(ras)
}
\arguments{
\item{ras}{raster}
}
\value{
array
}
\description{
Raster to RGBA array
}
