% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raster.R
\name{convert_raster_to_bytes}
\alias{convert_raster_to_bytes}
\title{Decode a raster}
\usage{
convert_raster_to_bytes(ras)
}
\arguments{
\item{ras}{raster}
}
\value{
raw vector
}
\description{
Decode a raster
}
