% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec-rasterGrob.R
\name{convert_rasterGrob_to_device_raster}
\alias{convert_rasterGrob_to_device_raster}
\title{Convert a rasterGrob into a device raster to mimic a graphics device}
\usage{
convert_rasterGrob_to_device_raster(raster_grob)
}
\arguments{
\item{raster_grob}{rasterGrob}
}
\value{
A \code{device-raster}. An integer vector, where each 32 bit integer
represents a packed colour value in ABGR format (which 8 bits per colour).
}
\description{
This function is needed to simulate the behaviour of R graphics devices.
}
\details{
When a rasterGrob is passed to a graphics device, it encodes the raster as a
\code{device-raster} which is an integer array containing packed colour values
in ABGR format.  This integer vector is then delivered to the graphics device
to display.
}
