% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec-bytes.R
\name{decode_bytes_to_robj}
\alias{decode_bytes_to_robj}
\title{Unserialize bytes into an R object}
\usage{
decode_bytes_to_robj(bytes)
}
\arguments{
\item{bytes}{vector of raw values}
}
\value{
the decoded R object if one is found, otherwise \code{NULL}
}
\description{
If no object is detected in the bytes, then \code{NULL} is returned
}
