% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec-device-raster.R
\name{decode_device_raster_to_bytes}
\alias{decode_device_raster_to_bytes}
\title{Convert a device raster to a raw vector}
\usage{
decode_device_raster_to_bytes(device_raster)
}
\arguments{
\item{device_raster}{a raster object passed from R graphics system to
the graphics device. It is an integer vector with packed ABGR values}
}
\description{
Convert a device raster to a raw vector
}
