% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec-device-raster.R
\name{decode_device_raster_to_robj}
\alias{decode_device_raster_to_robj}
\title{Extract a R object encoded in a device-raster}
\usage{
decode_device_raster_to_robj(device_raster)
}
\arguments{
\item{device_raster}{a raster object passed from R grpahics system to
the graphics device. It is an integer vector with packed ABGR values}
}
\value{
R object if there is one encoded in the device-raster, otherwise NULL
}
\description{
Return NULL of device-raster does not contain an object
}
