% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec-png.R
\name{decode_png_to_robj}
\alias{decode_png_to_robj}
\title{Decode a PNG file to an R object}
\usage{
decode_png_to_robj(filename)
}
\arguments{
\item{filename}{PNG filename}
}
\value{
If PNG contains an encode object then it is decoded and returned,
        otherwise NULL
}
\description{
Decode a PNG file to an R object
}
