% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec-raster.R
\name{decode_raster_to_robj}
\alias{decode_raster_to_robj}
\title{Raster to robject}
\usage{
decode_raster_to_robj(ras)
}
\arguments{
\item{ras}{raster}
}
\value{
R object if one is encoded in the raster, otherwsie NULL
}
\description{
Return NULL if the raster does not contain an encoded R object
}
