% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec-rasterGrob.R
\name{decode_rasterGrob_to_robj}
\alias{decode_rasterGrob_to_robj}
\title{Decode a rasterGrob to an R object}
\usage{
decode_rasterGrob_to_robj(raster_grob)
}
\arguments{
\item{raster_grob}{a \code{grid::rasterGrob()} object}
}
\description{
Decode a rasterGrob to an R object
}
