% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec-png.R
\name{encode_robj_to_png}
\alias{encode_robj_to_png}
\title{Encode an R object as a PNG file}
\usage{
encode_robj_to_png(robj, filename, alpha = 0)
}
\arguments{
\item{robj}{R object}

\item{filename}{PNG filename}

\item{alpha}{transparency. default: 0. Valid range [0, 1]}
}
\description{
Encode an R object as a PNG file
}
