% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec-rasterGrob.R
\name{encode_robj_to_rasterGrob}
\alias{encode_robj_to_rasterGrob}
\title{Encode an R object as a rasterGrob}
\usage{
encode_robj_to_rasterGrob(robj, alpha = 0, ...)
}
\arguments{
\item{robj}{Any R object}

\item{alpha}{alpha channel value. default: 0, i.e. raster is invisible.
This option  may be useful in debugging
i.e. set to non-zero value so raster is visible.  Valid range [0, 1]}

\item{...}{other arguments passed to \code{grid::rasterGrob()}}
}
\description{
The object is serialized and compressed and put into the RGB elements of
an RGBA array that is wrapped in a \code{rasterGrob}
}
