% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec-device-raster.R
\name{int32_to_raw}
\alias{int32_to_raw}
\title{Convert integer vector into raw vector (unpacking 4-raw-values per integer)}
\usage{
int32_to_raw(int32_vec, endian = c("little", "big"), ...)
}
\arguments{
\item{int32_vec}{vector of raw values}

\item{endian}{little or big?. default: 'litle'}

\item{...}{other arguments ignored}
}
\description{
Convert integer vector into raw vector (unpacking 4-raw-values per integer)
}
