% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codec-bytes.R
\name{encode_robj_to_bytes}
\alias{encode_robj_to_bytes}
\title{Encode an R object into a vector of raw bytes}
\usage{
encode_robj_to_bytes(robj)
}
\arguments{
\item{robj}{any object which can be handled with \code{base::serialize}}
}
\value{
Vector of raw bytes
}
\description{
If \code{zstdlite} is installed, then the serialized representation is
compressed using \code{zstd} compression.
}
